+++ internal/cli/plugin/list.go
@@ -35,6 +35,7 @@ func newListCmd(stdout, stderr io.Writer) *cobra.Command {
 }
 
 func runList(stdout, _ io.Writer, sourceFilter string) error {
+	cat := i18n.New(i18n.Detect())
 	if sourceFilter != "" &&
 		sourceFilter != plugin.SourceEmbedded &&
 		sourceFilter != plugin.SourceUser &&
@@ -51,7 +52,7 @@ func runList(stdout, _ io.Writer, sourceFilter string) error {
 	ids := slices.Sorted(maps.Keys(sources))
 
 	tw := tabwriter.NewWriter(stdout, 0, 0, 2, ' ', 0)
-	fmt.Fprintln(tw, "ID\tSOURCE\tDEFAULT\tDESCRIPTION\tURL")
+	fmt.Fprintln(tw, cat.Msg("plugin_list_header"))
 	for _, id := range ids {
 		src := sources[id]
 		if sourceFilter != "" && src != sourceFilter {
@@ -59,12 +60,12 @@ func runList(stdout, _ io.Writer, sourceFilter string) error {
 		}
 		p, lerr := loadPluginFromLayer(layered, id)
 		if lerr != nil {
-			fmt.Fprintf(tw, "%s\t%s\t?\t<load failed: %v>\t\n", id, src, lerr)
+			fmt.Fprintf(tw, "%s\t%s\t?\t%s\t\n", id, src, cat.Msg("plugin_list_load_failed", lerr))
 			continue
 		}
-		def := "no"
+		def := cat.Msg("plugin_list_no")
 		if p.Metadata.Default {
-			def = "yes"
+			def = cat.Msg("plugin_list_yes")
 		}
 		fmt.Fprintf(tw, "%s\t%s\t%s\t%s\t%s\n", id, src, def, p.Metadata.Description, p.Metadata.URL)
 	}